Fixed issue where plugins don't layout around the player if resizing=false.
Also added getPluginName and getPluginConfig to the View, which allows plugins to get their own configuration without knowing what flashvar was used to load them.  This feature is not available for plugins which are built into the player, such as controlbar and playlist.

Finally, added a RandomColor plugin to demonstrate the plugin resizing and getPluginConfig features.

git-svn-id: http://developer.longtailvideo.com/svn/trunk@108 6bf506d4-4edb-4b42-8bf1-84b36a5c7d31